chore: Update Contact and PhoneBook classes
The Contact class now includes methods to get and set the index, first name, last name, nickname, phone number, and darkest secret. The PhoneBook class has been updated to include a counter for the number of contacts and a variable to keep track of the index of the next contact to replace when the phone book is full. The addContact method has been modified to handle adding new contacts to the phone book. The printContact method has been updated to display the index, first name, last name, and nickname of each contact.